+++ tomcat/src/main/java/com/techcourse/controller/LoginController.java
@@ -9,8 +9,8 @@ import java.util.HashMap;
 import java.util.UUID;
 import org.apache.coyote.http11.FileReader;
 import org.apache.coyote.http11.request.HttpRequest;
-import org.apache.coyote.http11.request.Session;
-import org.apache.coyote.http11.request.SessionManager;
+import org.apache.coyote.http11.session.Session;
+import org.apache.coyote.http11.session.SessionManager;
 import org.apache.coyote.http11.response.HttpResponse;
 import org.apache.coyote.http11.response.HttpResponseBody;
 import org.apache.coyote.http11.response.HttpResponseHeaders;
