+++ tomcat/src/main/java/org/apache/coyote/http11/session/Session.java
@@ -1,4 +1,4 @@
-package org.apache.coyote.http11.request;
+package org.apache.coyote.http11.session;
 
 import java.util.HashMap;
 import java.util.Map;
